refactor: split AdminDataViewer into tab components and hooks (A2)
Mechanical relocation of the ~2,100-line AdminDataViewer.tsx:

- components/admin/AdminApp.tsx: shell (header, summary cards, tab nav,
  cross-tab "view AI logs for game" handoff)
- components/admin/utils.ts: pure helpers moved verbatim (turn grouping,
  symptom counting, formatters, copy-bundle builder, charge averages)
- components/admin/tabs/*: one component per tab, each owning its local
  state (selected log, expanded turns, selected playback, deck selection)
- hooks/useAdminData.ts: query hooks with identical keys/intervals
- hooks/useSimulationRuns.ts: simulation queries + start/cancel mutations;
  run-status polling converted from manual setInterval to useQuery with a
  refetchInterval that stops on terminal status (same 3s cadence)

New tests: utils unit tests, per-tab smoke renders, polling-stops hook test.
