+++ frontend/src/admin.tsx
@@ -6,7 +6,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import AdminDataViewer from './components/AdminDataViewer';
+import AdminApp from './components/admin/AdminApp';
 import './index.css';
 
 const queryClient = new QueryClient({
@@ -21,7 +21,7 @@ const queryClient = new QueryClient({
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
     <QueryClientProvider client={queryClient}>
-      <AdminDataViewer />
+      <AdminApp />
     </QueryClientProvider>
   </React.StrictMode>
 );
